youlan_kids_django/requirements.txt
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
requests==2.27.1
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0